classification_model/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.110.0
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
idna==3.6
iniconfig==2.0.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pillow==10.3.0
pluggy==1.4.0
protobuf==4.25.3
psycopg2-binary
pyasn1==0.5.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rich==13.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
termcolor==2.4.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
Werkzeug==3.0.2
wrapt==1.14.0
nltk
beautifulsoup4
scikit-learn
python-dotenv
boto3
